Updating test cases
