+++ tests/.gitignore
@@ -1,2 +1,3 @@
 cypress/videos/
 cypress/screenshots/
+cypress/reports/
