--- tests/cypress/downloads/swagger.null
@@ -1,1 +1,0 @@
-null
