--- tests/cypress/downloads/swagger.yaml
@@ -1,105 +1,0 @@
-openapi: 3.0.1
-info:
-  title: wordy-release
-  version: 1.0.0
-servers:
-  - url: /
-security:
-  - default: []
-paths:
-  /*:
-    get:
-      responses:
-        '200':
-          description: OK
-      security:
-        - default: []
-      x-auth-type: Application & Application User
-      x-throttling-tier: Unlimited
-      x-wso2-application-security:
-        security-types:
-          - oauth2
-        optional: false
-    put:
-      responses:
-        '200':
-          description: OK
-      security:
-        - default: []
-      x-auth-type: Application & Application User
-      x-throttling-tier: Unlimited
-      x-wso2-application-security:
-        security-types:
-          - oauth2
-        optional: false
-    post:
-      responses:
-        '200':
-          description: OK
-      security:
-        - default: []
-      x-auth-type: Application & Application User
-      x-throttling-tier: Unlimited
-      x-wso2-application-security:
-        security-types:
-          - oauth2
-        optional: false
-    delete:
-      responses:
-        '200':
-          description: OK
-      security:
-        - default: []
-      x-auth-type: Application & Application User
-      x-throttling-tier: Unlimited
-      x-wso2-application-security:
-        security-types:
-          - oauth2
-        optional: false
-    patch:
-      responses:
-        '200':
-          description: OK
-      security:
-        - default: []
-      x-auth-type: Application & Application User
-      x-throttling-tier: Unlimited
-      x-wso2-application-security:
-        security-types:
-          - oauth2
-        optional: false
-components:
-  securitySchemes:
-    default:
-      type: oauth2
-      flows:
-        implicit:
-          authorizationUrl: 'https://test.com'
-          scopes: {}
-x-wso2-auth-header: Authorization
-x-wso2-cors:
-  corsConfigurationEnabled: false
-  accessControlAllowOrigins:
-    - '*'
-  accessControlAllowCredentials: false
-  accessControlAllowHeaders:
-    - authorization
-    - Access-Control-Allow-Origin
-    - Content-Type
-    - SOAPAction
-    - apikey
-    - Internal-Key
-  accessControlAllowMethods:
-    - GET
-    - PUT
-    - POST
-    - DELETE
-    - PATCH
-    - OPTIONS
-x-wso2-basePath: /wordy_release/1.0.0
-x-wso2-transports:
-  - http
-  - https
-x-wso2-response-cache:
-  enabled: false
-  cacheTimeoutInSeconds: 300
