+++ tests/cypress/integration/devportal/001-applications/03-generate-api-keys.spec.js
@@ -17,18 +17,9 @@
 import Utils from "@support/utils";
 
 describe("Application tests", () => {
-    const appName = 'keygenapplication' + Math.floor(Date.now() / 1000);
+    const { developer, password } = Utils.getUserInfo();
+    const appName = Utils.generateName();
     const appDescription = 'Key gen application description';
-    const developer = 'developer';
-    const password = 'test123';
-    const carbonUsername = 'admin';
-    const carbonPassword = 'admin';
-
-    before(function () {
-        cy.carbonLogin(carbonUsername, carbonPassword);
-        cy.addNewUser(developer, ['Internal/subscriber', 'Internal/everyone'], password);
-    })
-
     const checkIfKeyExists = () => {
         // Check if the key exists
         cy.get('#access-token', { timeout: 30000 });
@@ -37,21 +28,7 @@ describe("Application tests", () => {
     }
     it.only("Generate API Keys", () => {
         cy.loginToDevportal(developer, password);
-        cy.visit(`${Utils.getAppOrigin()}/devportal/applications/create?tenant=carbon.super`);
-
-        // Filling the form
-        cy.get('#application-name')
-            .dblclick()
-            .type(appName);
-        cy.get('#application-description')
-            .click()
-            .type('{backspace}')
-            .type(appDescription);
-        cy.get('#itest-application-create-save').click();
-
-        // Checking the app name exists in the overview page.
-        cy.url().should('contain', '/overview');
-        cy.get('#itest-info-bar-application-name').contains(appName).should('exist');
+        cy.createApp(appName, appDescription);
 
         // Generating keys production
         cy.get('#production-keys-apikey').click();
@@ -82,13 +59,6 @@ describe("Application tests", () => {
     })
 
     after(() => {
-        cy.visit(`${Utils.getAppOrigin()}/devportal/applications?tenant=carbon.super`);
-        cy.get(`#delete-${appName}-btn`, { timeout: 30000 });
-        cy.get(`#delete-${appName}-btn`).click();
-        cy.get(`#itest-confirm-application-delete`).click();
-
-        // delete developer
-        cy.visit(`${Utils.getAppOrigin()}/carbon/user/user-mgt.jsp`);
-        cy.deleteUser(developer);
+        cy.deleteApp(appName);
     })
 })
