+++ tests/cypress/integration/devportal/002-subscriptions/00-subscribe-unsubscribe-to-app-from-api-details.spec.js
@@ -19,87 +19,72 @@
 import Utils from "@support/utils";
 
 describe("Subscribe unsubscribe to application from api details page", () => {
-    const appName = 'subscribeapp' + Math.floor(Date.now() / 1000);
-    const appDescription = 'app description';
-    const developer = 'developer';
-    const publisher = 'publisher';
-    const password = 'test123';
-    const carbonUsername = 'admin';
-    const carbonPassword = 'admin';
-    const apiName = `anonymous${Math.floor(Math.random() * (100000 - 1 + 1) + 1)}`;
-    const apiVersion = '2.0.0';
-    const apiContext = `anonymous${Math.floor(Math.random() * (100000 - 1 + 1) + 1)}`;
+    const { publisher, developer, password, tenantUser, tenant, } = Utils.getUserInfo();
 
-    before(function(){
-        cy.carbonLogin(carbonUsername, carbonPassword);
-        cy.addNewUser(developer, ['Internal/subscriber', 'Internal/everyone'], password);
-        cy.addNewUser(publisher, ['Internal/publisher', 'Internal/creator', 'Internal/everyone'], password);
-    })
-    
+    const apiVersion = '2.0.0';
+    const apiName = Utils.generateName();
+    const apiContext = apiName;
+    let testApiId;
+    const appName = Utils.generateName();
+    const appDescription = 'app description';
+
     it.only("Subscribe and unsubscribe to API from api details page", () => {
 
         cy.loginToPublisher(publisher, password);
-        cy.createAndPublishAPIByRestAPIDesign(apiName, apiVersion, apiContext);
-        cy.logoutFromPublisher();
-        cy.loginToDevportal(developer, password);
-        cy.createApp(appName, appDescription);
-        cy.visit(`${Utils.getAppOrigin()}/devportal/apis?tenant=carbon.super`);
-        cy.url().should('contain', '/apis?tenant=carbon.super');
-         // After publishing the api appears in devportal with a delay.
-        // We need to keep refresing and look for the api in the listing page
-        // following waitUntilApiExists function does that recursively.
-        let remainingAttempts = 30;
 
-        function waitUntilApiExists() {
-            let $apis = Cypress.$(`[title="${apiName}"]`);
-            if ($apis.length) {
-                // At least one with api name was found.
-                // Return a jQuery object.
-                return $apis;
-            }
-
-            if (--remainingAttempts) {
-                cy.log('Table not found yet. Remaining attempts: ' + remainingAttempts);
-
-                // Requesting the page to reload (F5)
-                cy.reload();
-
-                // Wait a second for the server to respond and the DOM to be present.
-                return cy.wait(4000).then(() => {
-                    return waitUntilApiExists();
+        Utils.addAPIWithEndpoints({ name: apiName, version: apiVersion, context: apiContext }).then((apiId) => {
+            testApiId = apiId;
+            Utils.publishAPI(apiId).then(() => {
+                cy.logoutFromPublisher();
+                cy.loginToDevportal(developer, password);
+                cy.createApp(appName, appDescription);
+                cy.visit(`${Utils.getAppOrigin()}/devportal/apis?tenant=carbon.super`);
+                cy.url().should('contain', '/apis?tenant=carbon.super');
+                 // After publishing the api appears in devportal with a delay.
+                // We need to keep refresing and look for the api in the listing page
+                // following waitUntilApiExists function does that recursively.
+                let remainingAttempts = 30;
+        
+                function waitUntilApiExists() {
+                    let $apis = Cypress.$(`[title="${apiName}"]`);
+                    if ($apis.length) {
+                        // At least one with api name was found.
+                        // Return a jQuery object.
+                        return $apis;
+                    }
+        
+                    if (--remainingAttempts) {
+                        cy.log('Table not found yet. Remaining attempts: ' + remainingAttempts);
+        
+                        // Requesting the page to reload (F5)
+                        cy.reload();
+        
+                        // Wait a second for the server to respond and the DOM to be present.
+                        return cy.wait(4000).then(() => {
+                            return waitUntilApiExists();
+                        });
+                    }
+                    throw Error('Table was not found.');
+                }
+        
+                waitUntilApiExists().then($apis => {
+                    cy.log('apis: ' + $apis.text());
+                    cy.get(`[title="${apiName}"]`, { timeout: 30000 });
+                    cy.get(`[title="${apiName}"]`).click();
+                    cy.get('#left-menu-credentials').click();
+            
+                    // Click and select the new application
+                    cy.get('#application-subscribe').click();
+                    cy.get(`.MuiAutocomplete-popper li`).contains(appName).click();
+                    cy.get(`#subscribe-to-api-btn`).click();
+                    cy.get(`#subscription-table td`).contains(appName).should('exist');
                 });
-            }
-            throw Error('Table was not found.');
-        }
-
-        waitUntilApiExists().then($apis => {
-            cy.log('apis: ' + $apis.text());
-            cy.get(`[title="${apiName}"]`, { timeout: 30000 });
-            cy.get(`[title="${apiName}"]`).click();
-            cy.get('#left-menu-credentials').click();
-    
-            // Click and select the new application
-            cy.get('#application-subscribe').click();
-            cy.get(`.MuiAutocomplete-popper li`).contains(appName).click();
-            cy.get(`#subscribe-to-api-btn`).click();
-            cy.get(`#subscription-table td`).contains(appName).should('exist');
+            });
         });
-       
     })
 
     after(() => {
-        cy.get(`#${appName}-UN`).click();
-        cy.visit(`${Utils.getAppOrigin()}/devportal/applications?tenant=carbon.super`);
-        cy.get(`#delete-${appName}-btn`, {timeout: 30000});
-        cy.get(`#delete-${appName}-btn`).click();
-        cy.get(`#itest-confirm-application-delete`).click();
-        cy.logoutFromDevportal();
-        cy.loginToPublisher(publisher, password);
-        cy.deleteApi(apiName, apiVersion);
-
-        // delete developer
-        cy.visit(`${Utils.getAppOrigin()}/carbon/user/user-mgt.jsp`);
-        cy.deleteUser(developer);
-        cy.deleteUser(publisher);
+        cy.deleteApp(appName);
+        Utils.deleteAPI(testApiId);
     })
 })
