+++ tests/cypress/integration/publisher/009-api-definition/01-download-swagger.spec.skip.js
