+++ tests/cypress/integration/publisher/011-lifecycle/02-deploy-as-prototype.spec.skip.js
