+++ tests/cypress/integration/publisher/013-api-product/02-create-a-new-revision-for-the-api-product-and-deploy.spec.skip.js
@@ -17,28 +17,16 @@
 import Utils from "@support/utils";
 
 describe("Mock the api response and test it", () => {
-    const publisher = 'publisher';
-    const password = 'test123';
-    const carbonUsername = 'admin';
-    const carbonPassword = 'admin';
-    const productName = 'petstoreProduct';
-    const apiName = 'SwaggerPetstore-OpenAPI30';
-    const apiVersion = '1.0.6';
+    const { publisher, password, } = Utils.getUserInfo();
+    const productName = Utils.generateName();
 
     before(function () {
-        cy.carbonLogin(carbonUsername, carbonPassword);
-        cy.addNewUser(publisher, ['Internal/publisher', 'Internal/creator', 'Internal/everyone'], password);
         cy.loginToPublisher(publisher, password);
     })
 
     it("Mock the api response and test it", () => {
-
-        cy.visit(`${Utils.getAppOrigin()}/publisher/apis`);
-        // select the option from the menu item
-        cy.get('#itest-rest-api-create-menu').click();
-        cy.get('#itest-id-landing-upload-oas').click();
+        cy.visit(`${Utils.getAppOrigin()}/publisher/apis/create/openapi`, { timeout: 30000 });
         cy.get('#open-api-file-select-radio').click();
-
         // upload the swagger
         cy.get('#browse-to-upload-btn').then(function () {
             const filepath = `api_artifacts/petstore-v3.json`
@@ -57,20 +45,16 @@ describe("Mock the api response and test it", () => {
             // finish the wizard
             cy.get('#open-api-create-btn').click();
 
-            cy.intercept({
-                method: "GET",
-                url: `**/apis/**`,
-                times: 1,
-              }).as('apiGet');
-            cy.wait('@apiGet', {timeout: 30000}).then((res) => {
-                
+            cy.intercept('**/apis/**').as('apiGet');
+            cy.wait('@apiGet', { timeout: 30000 }).then((res) => {
+
                 //Get the api id
                 const uuid = res.response.body.id;
 
                 // validate
                 cy.get('#itest-api-name-version', { timeout: 30000 });
                 cy.get('#itest-api-name-version').contains(version);
-                
+
                 // Go to api product create page
                 cy.visit(`${Utils.getAppOrigin()}/publisher/api-products/create`);
 
@@ -79,19 +63,14 @@ describe("Mock the api response and test it", () => {
                 cy.get('#context').type(productName);
                 cy.get('#itest-id-apiname-input').click();
 
+                cy.intercept('**/swagger').as('swaggerGet');
+
                 cy.get('#api-product-next-btn').click();
-
-                // Wait until the api is saved
-                cy.get('#resource-wrapper', { timeout: 30000 });
-                cy.get('#resource-wrapper').click();
-
-                cy.intercept("POST", `**/api-products`).as('apiProductsGet');
-
-                // add all resources
-                cy.get('#add-all-resources-btn').click();
-                cy.get('#create-api-product-btn').scrollIntoView().dblclick();
-
-                cy.wait('@apiProductsGet', {timeout: 30000}).then((res) => {
+                cy.wait('@swaggerGet', { timeout: 3000 }).then(() => {
+                    // add all resources
+                    cy.get('#add-all-resources-btn', { timeout: 30000 }).click();
+                    cy.get('span').contains('Define API Product').click();
+                    cy.get('#create-api-product-btn').scrollIntoView().dblclick();
 
                     cy.get('#itest-api-name-version', { timeout: 30000 });
                     cy.get('#itest-api-name-version').contains(productName);
@@ -100,33 +79,28 @@ describe("Mock the api response and test it", () => {
                     cy.get('#left-menu-itemdeployments').click();
 
                     // Deploying
-                    cy.get('#deploy-btn').click({"force":true});
+                    cy.get('#deploy-btn').click({ "force": true });
                     cy.get('#undeploy-btn').should('exist');
-                    cy.get('#undeploy-btn').click();
-                    cy.get('#revision-selector').should('exist');
 
-                    //Get the api product id
-                    const uuidProduct = res.response.body.id;
+                    // Going to lifecycle page
+                    cy.get('#left-menu-itemlifecycle').click();
 
-                    cy.log(uuid, uuidProduct);
+                    // Publishing
+                    cy.wait(2000);
+                    cy.get('[data-testid="Publish-btn"]').click();
 
-                    // Deleting the api and api product
-                    cy.visit(`${Utils.getAppOrigin()}/publisher/api-products/${uuidProduct}/overview`);
-                    cy.get('#itest-api-name-version', { timeout: 30000 });
-                    cy.get(`#itest-id-deleteapi-icon-button`).click({force: true});
+                    cy.get('button[data-testid="Demote to Created-btn"]').should('exist');
+
+                    cy.get(`#itest-id-deleteapi-icon-button`).click();
                     cy.get(`#itest-id-deleteconf`).click();
 
                     cy.visit(`${Utils.getAppOrigin()}/publisher/apis/${uuid}/overview`);
                     cy.get('#itest-api-name-version', { timeout: 30000 });
-                    cy.get(`#itest-id-deleteapi-icon-button`).click({force: true});
+                    cy.get(`#itest-id-deleteapi-icon-button`).click();
                     cy.get(`#itest-id-deleteconf`).click();
                 })
-            });    
+
+            });
         });
     });
-
-    after(function () {
-        cy.visit(`${Utils.getAppOrigin()}/carbon/user/user-mgt.jsp`);
-        cy.deleteUser(publisher);
-    })
 })
